Added support for older releases

+++ install.sh
@@ -230,14 +230,18 @@ install_x-ui() {
         rc-update add x-ui
         rc-service x-ui start
     else
-        case "${release}" in
-        ubuntu | debian | armbian)
-            cp -f x-ui.service.debian /etc/systemd/system/
-            ;;
-        *)
-            cp -f x-ui.service.rhel /etc/systemd/system/
-        ;;
-        esac
+        if [ -f "x-ui.service" ]; then
+            cp -f x-ui.service /etc/systemd/system/
+        else
+            case "${release}" in
+                ubuntu | debian | armbian)
+                    cp -f x-ui.service.debian /etc/systemd/system/
+                ;;
+                *)
+                    cp -f x-ui.service.rhel /etc/systemd/system/
+                ;;
+            esac
+        fi
         systemctl daemon-reload
         systemctl enable x-ui
         systemctl start x-ui
