+++ update.sh
@@ -227,15 +227,21 @@ update_x-ui() {
         rc-update add x-ui >/dev/null 2>&1
         rc-service x-ui start >/dev/null 2>&1
     else
-        echo -e "${green}Installing systemd unit...${plain}"
-        case "${release}" in
-        ubuntu | debian | armbian)
-            cp -f x-ui.service.debian /etc/systemd/system/ >/dev/null 2>&1
-            ;;
-        *)
-            cp -f x-ui.service.rhel /etc/systemd/system/ >/dev/null 2>&1
-        ;;
-        esac
+        if [ -f "x-ui.service" ]; then
+            echo -e "${green}Installing systemd unit...${plain}"
+            cp -f x-ui.service /etc/systemd/system/ >/dev/null 2>&1
+        else
+            case "${release}" in
+                ubuntu | debian | armbian)
+                    echo -e "${green}Installing debian-like systemd unit...${plain}"
+                    cp -f x-ui.service.debian /etc/systemd/system/ >/dev/null 2>&1
+                ;;
+                *)
+                    echo -e "${green}Installing rhel-like systemd unit...${plain}"
+                    cp -f x-ui.service.rhel /etc/systemd/system/ >/dev/null 2>&1
+                ;;
+            esac
+        fi
         chown root:root /etc/systemd/system/x-ui.service >/dev/null 2>&1
         systemctl daemon-reload >/dev/null 2>&1
         systemctl enable x-ui >/dev/null 2>&1
